requirements.txt
streamlit==1.31.1
pdfminer.six==20221105
python-docx==1.1.0
google-generativeai==0.4.1
